detection_backend/requirements.txt
confluent_kafka==2.12.0
facenet_pytorch==2.6.0
fastapi==0.118.3
insightface==0.7.3
loguru==0.7.3
opencv-contrib-python==4.11.0.86
Pillow==10.2.0
pydantic==2.12.0
pymongo==4.15.3
python-dotenv==1.1.1
python_dateutil==2.9.0.post0
PyYAML==6.0.2
qdrant_client==1.15.1
Requests==2.32.5
scikit-image==0.25.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
uvicorn==0.37.0
numpy==1.26.4
